requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# PDF Processing
pdfplumber==0.10.3
PyMuPDF==1.23.8

# NLP & AI
spacy==3.7.2
sentence-transformers==2.3.1
openai==1.10.0
google-generativeai==0.3.2

# Data Processing
numpy==1.26.3
scikit-learn==1.4.0

# Environment & Utils
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Email validation
email-validator==2.1.0
